fix: sql v16 적용 및 백엔드 수정
WatchRecord.rating을 Double에서 BigDecimal(DECIMAL(3,1))로, note 필드/review 컬럼을
privateReview/private_review로 통일했다. movie.vote_average와 타입을 맞춰 4-8의 대중
평점 비교에서 DECIMAL↔DOUBLE 혼합 연산 오차를 없애고, 컬럼명과 필드명이 갈려 있던
드리프트(isRepresentative 때와 같은 함정)를 닫았다.

- WatchRecord 엔티티: rating 타입, note→privateReview, validateRating() 범위
  0.0~10.0 → 1.0~10.0 정정
- WatchRecordCreateRequest/UpdateRequest/Response, WatchRecordService,
  UserMovieListItemResponse(컴파일 유지를 위해 함께 수정) 반영
- jpa-entity-spec.md/service-layer-spec.md/controller-layer-spec.md/CLAUDE.md 갱신,
  v16-delta.sql 추가

+++ src/main/java/com/project/cinemory/domain/watch/dto/UserMovieListItemResponse.java
@@ -6,6 +6,7 @@ import com.project.cinemory.domain.movie.entity.Movie;
 import com.project.cinemory.domain.watch.entity.WatchRecord;
 import com.project.cinemory.domain.watch.entity.WatchType;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -17,7 +18,7 @@ public record UserMovieListItemResponse(
         List<GenreResponse> genres,
         List<CountryResponse> countries,
         LocalDate watchDate,
-        Double rating,
+        BigDecimal rating,
         WatchType watchType
 ) {
